rename Tooltip.tsx to TimeSeriesTooltip, export all sub components
