+++ ui/components/src/TimeSeriesTooltip/SeriesInfo.tsx
