+++ ui/components/src/TimeSeriesTooltip/SeriesLabelsStack.tsx
